+++ example/lib/screens/screens.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:stripe_example/screens/address_sheet/address_sheet.dart';
 import 'package:stripe_example/screens/customer_sheet/customer_sheet_screen.dart';
 import 'package:stripe_example/screens/others/can_add_to_wallet_screen.dart';
+import 'package:stripe_example/screens/payment_sheet/embedded/embedded_payment_element_screen.dart';
 import 'package:stripe_example/screens/payment_sheet/express_checkout/express_checkout_element.dart';
 import 'package:stripe_example/screens/payment_sheet/payment_element/payment_element.dart';
 import 'package:stripe_example/screens/payment_sheet/payment_sheet_deffered_screen.dart';
@@ -125,6 +126,11 @@ class Example extends StatelessWidget {
         Example(
           title: 'Custom Flow',
           builder: (context) => PaymentSheetScreenWithCustomFlow(),
+          platformsSupported: [DevicePlatform.android, DevicePlatform.ios],
+        ),
+        Example(
+          title: 'Embedded Payment Element',
+          builder: (context) => EmbeddedPaymentElementScreen(),
           platformsSupported: [DevicePlatform.android, DevicePlatform.ios],
         ),
         Example(